Benchmark epigenomic profiling methods using motif enrichment

This function compares different epigenomic datasets using motif enrichment as the key metric. The output is an easy-to-interpret HTML document with the results. The report contains three main sections: (1) General Metrics on peak and alignment files (if provided), (2) Known Motif Enrichment Analysis and (3) De-novo Motif Enrichment Analysis.

MotifPeeker(
  peak_files,
  reference_index = 1,
  alignment_files = NULL,
  labels = NULL,
  exp_type = NULL,
  genome_build,
  motif_files = NULL,
  motif_labels = NULL,
  cell_counts = NULL,
  denovo_motif_discovery = TRUE,
  denovo_motifs = 3,
  motif_db = NULL,
  download_buttons = TRUE,
  output_dir = tempdir(),
  display = NULL,
  use_cache = TRUE,
  ncpus = 1,
  quiet = FALSE,
  debug = FALSE,
  verbose = FALSE
)

Arguments

peak_files

A character vector of path to peak files, or a vector of GRanges objects generated using read_peak_file. Currently, peak files from the following peak-calling tools are supported:

MACS2: .narrowPeak files

SEACR: .bed files

reference_index

An integer specifying the index of the peak file to use as the reference dataset for comparison. Indexing starts from 1. (default = 1)

alignment_files

A character vector of path to alignment files, or a vector of BamFile objects. (optional) Alignment files are used to calculate read-related metrics like FRiP score.

labels

A character vector of labels for each peak file. (optional) If not provided, the peak file names will be used as labels.

exp_type

A character vector of experimental types for each peak file. (optional) Useful for comparison of different methods. If not provided, all datasets will be classified as "unknown" experiment types in the report. Supported experimental types are:

chipseq: ChIP-seq data

tipseq: TIP-seq data

cuttag: CUT&Tag data

cutrun: CUT&Run data

exp_type is used only for labelling. It does not affect the analysis. You can also input custom strings. Datasets will be grouped as long as they match their respective exp_type.

genome_build

A character string with the abbreviated genome build name, or a BSGenome object. At the moment, only hg38 and hg19 are supported as abbreviated input.

motif_files

A character vector of path to motif files, or a vector of universalmotif-class objects. (optional) Required to run Known Motif Enrichment Analysis.

motif_labels

A character vector of labels for each motif file. (optional) Only used if path to file names are passed in motif_files. If not provided, the motif file names will be used as labels.

cell_counts

An integer vector of experiment cell counts for each peak file. (optional) Creates additional comparisons based on cell counts.

denovo_motif_discovery

A logical indicating whether to perform de-novo motif discovery for the third section of the report. (default = TRUE)

denovo_motifs

An integer specifying the number of de-novo motifs to discover. (default = 3) Note that higher values take longer to compute.

motif_db

Path to .meme format file to use as reference database, or a list of universalmotif-class objects. (optional) Results from de-novo motif discovery are searched against this database to find similar motifs. If not provided, JASPAR CORE database will be used. NOTE: p-value estimates are inaccurate when the database has fewer than 50 entries.

download_buttons

A logical indicating whether to include download buttons for various files within the HTML report. (default = TRUE)

output_dir

A character string specifying the directory to save the output files. (default = tempdir()) A sub-directory with the output files will be created in this directory.

display

A character vector specifying the display mode for the HTML report once it is generated. (default = NULL) Options are:

"browser": Open the report in the default web browser.

"rstudio": Open the report in the RStudio Viewer.

NULL: Do not open the report.

use_cache

A logical indicating whether to use cached results from previous runs. (default = TRUE)

ncpus

An integer specifying the number of cores to use for parallel processing. (default = 1)

quiet

A logical indicating whether to suppress messages in the HTML report. (default = FALSE)

debug

A logical indicating whether to print debug messages in the HTML report. (default = FALSE)

verbose

A logical indicating whether to print verbose messages while running the function. (default = FALSE)

Value

Path to the output directory.

Details

The function can optionally call peaks using MACS3, but it is recommended that the peak files be supplied directly and produced using fine-tuned parameters.

Note

Running de-novo motif discovery is computationally expensive and can require from minutes to hours. denovo_motifs can widely affect the runtime (higher values take longer).