Get datatable for motif-enrichment of individual experiments.

Get datatable for motif-enrichment of individual experiments.

dt_enrichment_individual(
  result,
  enrichment_df,
  comparison_i,
  motif_i,
  reference_index = 1
)

Arguments

result

A list with the following elements:

peaks

A list of peak files generated using read_peak_file.

alignments

A list of alignment files.

exp_type

A character vector of experiment types.

exp_labels

A character vector of experiment labels.

read_count

A numeric vector of read counts.

peak_count

A numeric vector of peak counts.

enrichment_df

A data frame containing the motif enrichment results, produced using get_df_enrichment.

comparison_i

The index of the comparison dataset to plot.

motif_i

The index of the motif to plot.

reference_index

An integer specifying the index of the peak file to use as the reference dataset for comparison. Indexing starts from 1. (default = 1)

Value

A DT::datatable object with the peak motif enrichment data for the specified comparison_i and motif_i.

See also