Get parameters for BiocParallel

Get appropriate parameters for BiocParallel based on the number of workers specified. For less than 10 workers, the function returns a MulticoreParam object. For 10 or more cores, the function returns a SnowParam object. Since Windows supports neither, the function returns a SerialParam object. As a result, Windows users do not benefit from parallel processing.

Usage

get_bpparam(
  workers,
  progressbar = workers > 1,
  force_snowparam = FALSE,
  verbose = FALSE
)

Arguments

workers

The number of workers to use for parallel processing.

progressbar

logical(1) Enable progress bar (based on plyr:::progress_text).

force_snowparam

A logical indicating whether to force the use of SnowParam object.

verbose

A logical indicating whether to print verbose messages while running the function. (default = FALSE)

Value

A BPPARAM object.

See also