Get dataframe with bootstrapped motif-summit distances

Wrapper for `MotifPeeker::bootstrap_distances` to get bootstrapped motif-summit distances for given peaks and motifs, generating a data.frame suitable for plots.

Usage

get_df_distances_bootstrapped(
  result,
  user_motifs,
  genome_build,
  samples_n = NULL,
  samples_len = NULL,
  out_dir = tempdir(),
  BPPARAM = BiocParallel::bpparam(),
  meme_path = NULL,
  verbose = FALSE
)

Arguments

result

A list with the following elements:

peaks

A list of peak files generated using read_peak_file.

alignments

A list of alignment files.

exp_type

A character vector of experiment types.

exp_labels

A character vector of experiment labels.

read_count

A numeric vector of read counts.

peak_count

A numeric vector of peak counts.

user_motifs

A list with the following elements:

motifs

A list of motif files.

motif_labels

A character vector of motif labels.

genome_build

A character string with the abbreviated genome build name, or a BSGenome object. Check check_genome_build details for genome builds which can be imported as abbreviated names.

samples_n

An integer specifying the number of bootstrap samples to generate. If NULL, it is set to 70% of the number of peaks.

samples_len

An integer specifying the number of peaks to sample in each bootstrap iteration. If NULL, it is set to 20 peaks.

out_dir

A character vector of output directory.

BPPARAM

A BiocParallelParam-class object enabling parallel execution. (default = SerialParam(), single-CPU run)

Following are two examples of how to set up parallel processing:

BPPARAM = BiocParallel::MulticoreParam(4): Uses 4 CPU cores for parallel processing.

library("BiocParallel") followed by register(MulticoreParam(4)) sets all subsequent BiocParallel functions to use 4 CPU cores. Motifpeeker() must be run with BPPARAM = BiocParallel::MulticoreParam().

IMPORTANT: For each worker, please ensure a minimum of 8GB of memory (RAM) is available as motif_discovery is memory-intensive.

meme_path

path to meme/bin/ (optional). Defaut: NULL, searches "MEME_PATH" environment variable or "meme_path" option for path to "meme/bin/".

verbose

A logical indicating whether to print verbose messages while running the function. (default = FALSE)

Value

A data.frame with the following columns:

exp_label

Experiment labels.

exp_type

Experiment types.

motif_indice

Motif indices.

bootstrap_iteration

Bootstrap iteration number.

distance

Mean of absolute distances between peak summit and motif.

See also

Examples

if (memes::meme_is_installed()) {
    peak <- system.file("extdata", "CTCF_ChIP_peaks.narrowPeak",
                package = "MotifPeeker") |>
                read_peak_file() |>
                sample(20)
    motif_MA1102.3 <- system.file("extdata", "motif_MA1102.3.jaspar",
                package = "MotifPeeker") |> read_motif_file()
    motif_MA1930.2 <- system.file("extdata", "motif_MA1930.2.jaspar",
                package = "MotifPeeker") |> read_motif_file()

    input <- list(
        peaks = peak,
        exp_type = "ChIP",
        exp_labels = "CTCF",
        read_count = 150,
        peak_count = 100
    )
    motifs <- list(
        motifs = list(motif_MA1930.2, motif_MA1102.3),
        motif_labels = list("MA1930.2", "MA1102.3")
    )
                
    if (requireNamespace("BSgenome.Hsapiens.UCSC.hg38")) {
        genome_build <-
            BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38
            
        distances_df_bootstrapped <- get_df_distances_bootstrapped(
            input,
            user_motifs = motifs,
            genome_build = genome_build,
            samples_n = NULL,
            samples_len = NULL,
            verbose = FALSE
        )
        print(distances_df_bootstrapped)
    }
}
#>     exp_label exp_type motif_indice bootstrap_iteration   distance
#> 1        CTCF     ChIP            1                   1   9.500000
#> 2        CTCF     ChIP            1                   2  25.500000
#> 3        CTCF     ChIP            1                   3  14.500000
#> 4        CTCF     ChIP            1                   4  28.666667
#> 5        CTCF     ChIP            1                   5  22.000000
#> 6        CTCF     ChIP            1                   6  21.400000
#> 7        CTCF     ChIP            1                   7  14.000000
#> 8        CTCF     ChIP            1                   8  23.000000
#> 9        CTCF     ChIP            1                   9  11.000000
#> 10       CTCF     ChIP            1                  10  20.500000
#> 11       CTCF     ChIP            1                  11   6.000000
#> 12       CTCF     ChIP            1                  12   6.666667
#> 13       CTCF     ChIP            1                  13  10.333333
#> 14       CTCF     ChIP            1                  14  20.800000
#> 15       CTCF     ChIP            1                  15  38.000000
#> 16       CTCF     ChIP            1                  16  19.250000
#> 17       CTCF     ChIP            1                  17  22.000000
#> 18       CTCF     ChIP            1                  18  23.333333
#> 19       CTCF     ChIP            1                  19  12.000000
#> 20       CTCF     ChIP            1                  20  14.250000
#> 21       CTCF     ChIP            1                  21  29.666667
#> 22       CTCF     ChIP            1                  22  19.000000
#> 23       CTCF     ChIP            1                  23  20.600000
#> 24       CTCF     ChIP            1                  24  27.000000
#> 25       CTCF     ChIP            1                  25  18.166667
#> 26       CTCF     ChIP            1                  26  23.200000
#> 27       CTCF     ChIP            1                  27  24.166667
#> 28       CTCF     ChIP            1                  28  25.285714
#> 29       CTCF     ChIP            1                  29  25.200000
#> 30       CTCF     ChIP            1                  30  15.600000
#> 31       CTCF     ChIP            1                  31   6.000000
#> 32       CTCF     ChIP            1                  32  25.833333
#> 33       CTCF     ChIP            1                  33  16.000000
#> 34       CTCF     ChIP            1                  34  19.000000
#> 35       CTCF     ChIP            1                  35  21.428571
#> 36       CTCF     ChIP            1                  36  22.166667
#> 37       CTCF     ChIP            1                  37  24.250000
#> 38       CTCF     ChIP            1                  38   6.000000
#> 39       CTCF     ChIP            1                  39  24.000000
#> 40       CTCF     ChIP            1                  40  46.000000
#> 41       CTCF     ChIP            1                  41  20.000000
#> 42       CTCF     ChIP            1                  42   6.666667
#> 43       CTCF     ChIP            1                  43  16.666667
#> 44       CTCF     ChIP            1                  44  18.166667
#> 45       CTCF     ChIP            1                  45  16.750000
#> 46       CTCF     ChIP            1                  46  13.500000
#> 47       CTCF     ChIP            1                  47  16.800000
#> 48       CTCF     ChIP            1                  48  19.333333
#> 49       CTCF     ChIP            1                  49  20.500000
#> 50       CTCF     ChIP            1                  50  21.833333
#> 51       CTCF     ChIP            1                  51  21.400000
#> 52       CTCF     ChIP            1                  52  20.500000
#> 53       CTCF     ChIP            1                  53   7.000000
#> 54       CTCF     ChIP            1                  54  21.800000
#> 55       CTCF     ChIP            1                  55  32.250000
#> 56       CTCF     ChIP            1                  56  29.666667
#> 57       CTCF     ChIP            1                  57  18.571429
#> 58       CTCF     ChIP            1                  58  24.333333
#> 59       CTCF     ChIP            1                  59  27.285714
#> 60       CTCF     ChIP            1                  60  14.500000
#> 61       CTCF     ChIP            1                  61  13.000000
#> 62       CTCF     ChIP            1                  62  24.166667
#> 63       CTCF     ChIP            1                  63  22.166667
#> 64       CTCF     ChIP            1                  64  12.000000
#> 65       CTCF     ChIP            1                  65  16.500000
#> 66       CTCF     ChIP            1                  66  16.500000
#> 67       CTCF     ChIP            1                  67  20.600000
#> 68       CTCF     ChIP            1                  68  29.750000
#> 69       CTCF     ChIP            1                  69  11.333333
#> 70       CTCF     ChIP            1                  70  16.250000
#> 71       CTCF     ChIP            1                  71  18.000000
#> 72       CTCF     ChIP            1                  72  22.166667
#> 73       CTCF     ChIP            1                  73  20.000000
#> 74       CTCF     ChIP            1                  74  14.250000
#> 75       CTCF     ChIP            1                  75  16.666667
#> 76       CTCF     ChIP            1                  76  13.750000
#> 77       CTCF     ChIP            1                  77  38.000000
#> 78       CTCF     ChIP            1                  78  18.285714
#> 79       CTCF     ChIP            1                  79  23.000000
#> 80       CTCF     ChIP            1                  80  18.166667
#> 81       CTCF     ChIP            1                  81  17.000000
#> 82       CTCF     ChIP            1                  82  20.800000
#> 83       CTCF     ChIP            1                  83  14.800000
#> 84       CTCF     ChIP            1                  84  17.500000
#> 85       CTCF     ChIP            1                  85  10.333333
#> 86       CTCF     ChIP            1                  86  20.600000
#> 87       CTCF     ChIP            1                  87  22.600000
#> 88       CTCF     ChIP            1                  88  26.500000
#> 89       CTCF     ChIP            1                  89  20.200000
#> 90       CTCF     ChIP            1                  90  20.800000
#> 91       CTCF     ChIP            1                  91  22.166667
#> 92       CTCF     ChIP            1                  92  19.000000
#> 93       CTCF     ChIP            1                  93  17.000000
#> 94       CTCF     ChIP            1                  94  20.000000
#> 95       CTCF     ChIP            1                  95  21.833333
#> 96       CTCF     ChIP            1                  96  18.285714
#> 97       CTCF     ChIP            1                  97   9.500000
#> 98       CTCF     ChIP            1                  98  16.500000
#> 99       CTCF     ChIP            1                  99   5.000000
#> 100      CTCF     ChIP            1                 100  20.000000
#> 101      CTCF     ChIP            2                   1 126.500000
#> 102      CTCF     ChIP            2                   2 126.500000
#> 103      CTCF     ChIP            2                   3  51.500000
#> 104      CTCF     ChIP            2                   4  51.500000
#> 105      CTCF     ChIP            2                   5  16.500000
#> 106      CTCF     ChIP            2                   6         NA
#> 107      CTCF     ChIP            2                   7  33.500000
#> 108      CTCF     ChIP            2                   8  16.500000
#> 109      CTCF     ChIP            2                   9  71.500000
#> 110      CTCF     ChIP            2                  10  99.500000
#> 111      CTCF     ChIP            2                  11         NA
#> 112      CTCF     ChIP            2                  12  71.500000
#> 113      CTCF     ChIP            2                  13  11.500000
#> 114      CTCF     ChIP            2                  14  47.000000
#> 115      CTCF     ChIP            2                  15  71.500000
#> 116      CTCF     ChIP            2                  16         NA
#> 117      CTCF     ChIP            2                  17  42.000000
#> 118      CTCF     ChIP            2                  18 126.500000
#> 119      CTCF     ChIP            2                  19  16.500000
#> 120      CTCF     ChIP            2                  20         NA
#> 121      CTCF     ChIP            2                  21  16.500000
#> 122      CTCF     ChIP            2                  22 126.500000
#> 123      CTCF     ChIP            2                  23         NA
#> 124      CTCF     ChIP            2                  24         NA
#> 125      CTCF     ChIP            2                  25  51.500000
#> 126      CTCF     ChIP            2                  26  71.500000
#> 127      CTCF     ChIP            2                  27  99.500000
#> 128      CTCF     ChIP            2                  28  44.500000
#> 129      CTCF     ChIP            2                  29  39.166667
#> 130      CTCF     ChIP            2                  30  16.500000
#> 131      CTCF     ChIP            2                  31  16.500000
#> 132      CTCF     ChIP            2                  32         NA
#> 133      CTCF     ChIP            2                  33         NA
#> 134      CTCF     ChIP            2                  34  16.500000
#> 135      CTCF     ChIP            2                  35  11.500000
#> 136      CTCF     ChIP            2                  36 126.500000
#> 137      CTCF     ChIP            2                  37  51.500000
#> 138      CTCF     ChIP            2                  38  42.000000
#> 139      CTCF     ChIP            2                  39 126.500000
#> 140      CTCF     ChIP            2                  40         NA
#> 141      CTCF     ChIP            2                  41 126.500000
#> 142      CTCF     ChIP            2                  42  71.500000
#> 143      CTCF     ChIP            2                  43  69.000000
#> 144      CTCF     ChIP            2                  44 126.500000
#> 145      CTCF     ChIP            2                  45 126.500000
#> 146      CTCF     ChIP            2                  46 126.500000
#> 147      CTCF     ChIP            2                  47  71.500000
#> 148      CTCF     ChIP            2                  48         NA
#> 149      CTCF     ChIP            2                  49  16.500000
#> 150      CTCF     ChIP            2                  50  56.750000
#> 151      CTCF     ChIP            2                  51         NA
#> 152      CTCF     ChIP            2                  52         NA
#> 153      CTCF     ChIP            2                  53  33.500000
#> 154      CTCF     ChIP            2                  54  69.000000
#> 155      CTCF     ChIP            2                  55 126.500000
#> 156      CTCF     ChIP            2                  56         NA
#> 157      CTCF     ChIP            2                  57  16.500000
#> 158      CTCF     ChIP            2                  58  71.500000
#> 159      CTCF     ChIP            2                  59 126.500000
#> 160      CTCF     ChIP            2                  60  14.000000
#> 161      CTCF     ChIP            2                  61 126.500000
#> 162      CTCF     ChIP            2                  62  16.500000
#> 163      CTCF     ChIP            2                  63  16.500000
#> 164      CTCF     ChIP            2                  64  51.500000
#> 165      CTCF     ChIP            2                  65         NA
#> 166      CTCF     ChIP            2                  66  69.000000
#> 167      CTCF     ChIP            2                  67 126.500000
#> 168      CTCF     ChIP            2                  68 126.500000
#> 169      CTCF     ChIP            2                  69 126.500000
#> 170      CTCF     ChIP            2                  70  51.500000
#> 171      CTCF     ChIP            2                  71  71.833333
#> 172      CTCF     ChIP            2                  72 126.500000
#> 173      CTCF     ChIP            2                  73  71.500000
#> 174      CTCF     ChIP            2                  74         NA
#> 175      CTCF     ChIP            2                  75 126.500000
#> 176      CTCF     ChIP            2                  76  16.500000
#> 177      CTCF     ChIP            2                  77  16.500000
#> 178      CTCF     ChIP            2                  78         NA
#> 179      CTCF     ChIP            2                  79         NA
#> 180      CTCF     ChIP            2                  80  69.000000
#> 181      CTCF     ChIP            2                  81         NA
#> 182      CTCF     ChIP            2                  82         NA
#> 183      CTCF     ChIP            2                  83  16.500000
#> 184      CTCF     ChIP            2                  84  51.500000
#> 185      CTCF     ChIP            2                  85  42.000000
#> 186      CTCF     ChIP            2                  86  44.500000
#> 187      CTCF     ChIP            2                  87         NA
#> 188      CTCF     ChIP            2                  88  11.500000
#> 189      CTCF     ChIP            2                  89  40.000000
#> 190      CTCF     ChIP            2                  90  56.750000
#> 191      CTCF     ChIP            2                  91  56.750000
#> 192      CTCF     ChIP            2                  92  56.750000
#> 193      CTCF     ChIP            2                  93  71.500000
#> 194      CTCF     ChIP            2                  94  71.500000
#> 195      CTCF     ChIP            2                  95  51.500000
#> 196      CTCF     ChIP            2                  96         NA
#> 197      CTCF     ChIP            2                  97         NA
#> 198      CTCF     ChIP            2                  98  44.500000
#> 199      CTCF     ChIP            2                  99  16.500000
#> 200      CTCF     ChIP            2                 100  51.500000