Print denovo motif enrichment datatable and download buttons for related files.

Print denovo motif enrichment datatable and download buttons for related files.

Usage

print_denovo_sections(
  motif_list,
  similar_motifs,
  segregated_peaks,
  indices,
  jaspar_link = FALSE,
  download_buttons = NULL
)

Arguments

motif_list

A list of motifs discovered by find_motifs, for one comparison pair.

similar_motifs

A list of similar motifs discovered using motif_similarity, for one comparison pair.

segregated_peaks

A list of peaks segregated by common and unique groups, for one comparison pair.

indices

A list of indices to print the datatable and download buttons for.

jaspar_link

A logical indicating whether to include a link to the JASPAR database for the motifs. Only set to TRUE if the motifs are in JASPAR format (example: "MA1930.1").

download_buttons

Embed download buttons generated using get_download_buttons. If set to NULL, no download buttons will be added.

Value

Null

See also